requirements.txt
flask
flask-cors
pillow